+++ CHANGELOG.md
@@ -41,3 +41,75 @@
 ## 3.0.5
 
 - Upgrade deprecated version of Android embedding
+
+## 3.1.0
+
+- Upgrade ios and android sdk versions
+
+## 3.1.1
+
+- Fix null userid in case of android device
+
+## 3.1.2
+
+- change fynouser.setinapp to fynosdk.registerinapp
+
+## 3.1.3
+
+- Removed images from the package
+
+## 3.1.4
+
+- Upgraded android sdk versions
+
+## 3.1.5
+
+- Upgraded android sdk versions
+
+## 3.1.6
+
+- Upgraded android sdk versions - table does not exist bug
+
+## 3.1.7
+
+- Upgraded android sdk versions - added debug logs
+
+## 3.2.0
+
+- updated android sdk versions and added versioning for iOS sdk
+
+## 3.2.1
+
+- added function to fetch push notification token
+
+## 3.3.0
+
+- logic to force update push token on first installation in new application (handled in kotlin and swift sdk)
+- function to update name
+
+## 3.3.1
+
+- push notification clicked listener
+
+## 3.4.0
+
+- added isFynoNotification and handleFynoNotification functions
+
+## 3.4.1
+
+- call renderFCMMessage inside CoroutineScope(Dispatchers.IO)
+
+## 3.4.2
+
+- updated android sdk versions - additional checks before API calls
+
+## 3.4.3
+
+- updated android sdk versions - include error handling for failed network requests
+
+## 3.4.4
+
+- updated ios sdk version - approach change of database processing
+
+## 3.5.0
+- updated android and ios sdk versions - reset and initialize flow approach change